# models/openAPI.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AudiMeta API",
        "version": "1.0.2",
        "description": "API to retrieve information about Book, Series and Search"
    },
    "servers": [
        {
            "url": "https://audimeta.de/",
            "description": "Public Instance"
        }
    ],
    "paths": {
...
        "/author/{asin}": {
            "get": {
                "tags": [
                    "author"
                ],
                "summary": "Get an author",
                "parameters": [
                    {
                        "name": "region",
                        "in": "query",
                        "description": "The region to search in. Can be a comma separated list of max 3 regions. Defaults to US",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "US",
                                "UK",
                                "AU",
                                "DE",
                                "FR",
                                "CA",
                                "IT",
                                "JP",
                                "IN",
                                "BR",
                                "MX",
                                "ES"
                            ]
                        }
                    },
                    {
                        "name": "asin",
                        "in": "path",
                        "description": "The asin of the author",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Author found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "asin": {
                                            "type": "string",
                                            "description": "The author asin. Can be the authors name!"
                                        },
                                        "region": {
                                            "type": "string",
                                            "enum": [
                                                "US",
                                                "UK",
                                                "AU",
                                                "DE",
                                                "FR",
                                                "CA",
                                                "IT",
                                                "JP",
                                                "IN",
                                                "BR",
                                                "MX",
                                                "ES"
                                            ],
                                            "description": "The language of the authors description"
                                        },
                                        "name": {
                                            "type": "string",
                                            "description": "The name of the author"
                                        },
                                        "description": {
                                            "type": "string",
                                            "description": "The description of the author"
                                        },
                                        "image": {
                                            "type": "string",
                                            "description": "The image of the author. 512x512 resolution"
                                        },
                                        "genres": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "asin": {
                                                        "type": "string",
                                                        "description": "The genre asin"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "description": "The name of the genre"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "Genres",
                                                            "Tags"
                                                        ],
                                                        "description": "The type of the genre"
                                                    }
                                                },
                                                "required": ["asin", "name", "type"]
                                            },
                                            "description": "The genres of the author"
                                        }
                                    },
                                    "required": ["asin", "region", "name"]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/author": {
            "get": {
                "tags": [
                    "author"
                ],
                "summary": "Search for authors",
                "parameters": [
                    {
                        "name": "region",
                        "in": "query",
                        "description": "The region to search in. Can be a comma separated list of max 3 regions. Defaults to US",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "US",
                                "UK",
                                "AU",
                                "DE",
                                "FR",
                                "CA",
                                "IT",
                                "JP",
                                "IN",
                                "BR",
                                "MX",
                                "ES"
                            ]
                        }
                    },
                    {
                        "name": "cache",
                        "in": "query",
                        "description": "boolean. Defaults to true if not set. If false then a query will be made to Audible instead of the cache. For endpoints that return multiple books this will only ignore the search cache. For endpoints that return a single book this will ignore the book cache and re-fetch the book from source",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the author",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authors found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "asin": {
                                                "type": "string",
                                                "description": "The author asin. Can be the authors name!"
                                            },
                                            "region": {
                                                "type": "string",
                                                "enum": [
                                                    "US",
                                                    "UK",
                                                    "AU",
                                                    "DE",
                                                    "FR",
                                                    "CA",
                                                    "IT",
                                                    "JP",
                                                    "IN",
                                                    "BR",
                                                    "MX",
                                                    "ES"
                                                ],
                                                "description": "The language of the authors description"
                                            },
                                            "name": {
                                                "type": "string",
                                                "description": "The name of the author"
                                            },
                                            "description": {
                                                "type": "string",
                                                "description": "The description of the author"
                                            },
                                            "image": {
                                                "type": "string",
                                                "description": "The image of the author. 512x512 resolution"
                                            },
                                            "genres": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "asin": {
                                                            "type": "string",
                                                            "description": "The genre asin"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "description": "The name of the genre"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "enum": [
                                                                "Genres",
                                                                "Tags"
                                                            ],
                                                            "description": "The type of the genre"
                                                        }
                                                    },
                                                    "required": ["asin", "name", "type"]
                                                },
                                                "description": "The genres of the author"
                                            }
                                        },
                                        "required": ["asin", "region", "name"]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}